- add tests for validators - add tests for exceptions
