+++ src/validation/pathvariablecollision/colliding-segment-node.ts
@@ -1,0 +1,6 @@
+import { ISegmentNode } from '../../parser';
+
+export interface CollidingSegmentNodes {
+  pathVariables: ISegmentNode[];
+  nonPathVariables: ISegmentNode[];
+}
