+++ src/validation/reservedroutesegment/reserved-route-segment.validator.spec.ts
@@ -1,0 +1,3 @@
+import { ISegmentNode } from '../../parser';
+
+describe('testing invalid url segment validator', () => {});
