+++ src/validation/reservedroutesegment/reserved-route.ts
@@ -1,0 +1,6 @@
+export const RESERVED_ROUTE_KEYWORD_REGEX = {
+  GET_NAME: /^get(-_)name$/i,
+  BUILD: /^build$/i,
+  WITH_PARAMS: /^with(-_)params$/i,
+  WITH_PARAM: /^with(-_)param$/i,
+};
